Add files via upload
In this DS Automation, We have loaded our prepared data and have dropped the percentage rate charge's variable, so it would fit the model we created after. We made a copy of our data, and so we could run our lab with the copy instead of the original. Then, we have used pycaret to find an ML algorithm that perform best on our data. We save the model on our disk. We have created a Python script/file/module with a function that takes a pandas dataframe as an input and returns the probability of churn for each row in the dataframe (We have used the new data, new_churn_data.csv).

Finally, we have tested our Python module and function with the new data, new_churn_data.csv